separated dbOptions and createResponse, added new errors
