+++ product-service/db/selectProductById.js
@@ -1,24 +1,5 @@
 import {Client} from 'pg';
-
-const {
-    AWS_PG_HOST,
-    AWS_PG_PORT,
-    AWS_PG_DATABASE,
-    AWS_PG_USERNAME,
-    AWS_PG_PASSWORD
-} = process.env;
-
-const dbOptions = {
-    host: AWS_PG_HOST,
-    port: AWS_PG_PORT,
-    database: AWS_PG_DATABASE,
-    user: AWS_PG_USERNAME,
-    password: AWS_PG_PASSWORD,
-    ssl: {
-        rejectUnauthorized: false
-    },
-    connectionTimeoutMillis: 10000
-};
+import dbOptions from './dbOptions';
 	
 export const selectProductById = async (productId) => {
     const client = new Client(dbOptions);
@@ -29,10 +10,17 @@ export const selectProductById = async (productId) => {
         const {rows: product} = await client.query(`
             SELECT id, title, description, price, count FROM public.products INNER JOIN public.stocks ON id = product_id WHERE id = '${productId}'
         `);
-        return {
-            status: 200,
-            product
-        };
+        if (product.length) {
+            return {
+                status: 200,
+                product
+            };
+        } else {
+            return {
+                status: 404,
+                error: {message: 'Product not found'}
+            };
+        }
     } catch (err) {
         return {
             status: 500,
